+++ ldm/modules/image_degradation/utils_image.py
@@ -85,7 +85,7 @@ def _get_paths_from_images(path):
 
 '''
 # --------------------------------------------
-# split large images into small images 
+# split large images into small images
 # --------------------------------------------
 '''
 
@@ -628,8 +628,8 @@ def calculate_psnr(img1, img2, border=0):
     img1 = img1[border:h-border, border:w-border]
     img2 = img2[border:h-border, border:w-border]
 
-    img1 = img1.astype(np.float64)
-    img2 = img2.astype(np.float64)
+    img1 = img1.astype(np.float32)
+    img2 = img2.astype(np.float32)
     mse = np.mean((img1 - img2)**2)
     if mse == 0:
         return float('inf')
@@ -670,8 +670,8 @@ def ssim(img1, img2):
     C1 = (0.01 * 255)**2
     C2 = (0.03 * 255)**2
 
-    img1 = img1.astype(np.float64)
-    img2 = img2.astype(np.float64)
+    img1 = img1.astype(np.float32)
+    img2 = img2.astype(np.float32)
     kernel = cv2.getGaussianKernel(11, 1.5)
     window = np.outer(kernel, kernel.transpose())
